pyproject.toml
[tool.poetry]
name = "stock-ainalyst"
version = "0.1.0"
description = ""
authors = ["Seungjae Park <astro.psj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.30.1"
opendartreader = "^0.2.3"
pandas = "^2.2.2"
psycopg = {extras = ["binary"], version = "^3.1.19"}
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.2"
requests = "^2.32.1"
translators = "^5.9.2"
yfinance = "^0.2.40"
pgvector = "^0.2.5"

[tool.black]
line-length = 200

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
